testing limit: I don't! I was just playing around with it, and didn't manage to fire off any character limit warnings. I'll try to push it to the edge haha
10:23
From the interwebz, the title limit is 72 char, and the message has no hard limits I don't! I was just playing around with it, and didn't manage to fire off any character limit warnings. I'll try to push it to the edge haha

10:23
From the interwebz, the title limit is 72 char, and the message has no hard limits I don't! I was just playing around with it, and didn't manage to fire off any character limit warnings. I'll try to push it to the edge haha

10:23
From the interwebz, the title limit is 72 char, and the message has no hard limits

description as well